requirements.txt
Django>=5.0,<5.1
psycopg2-binary>=2.9.9
Pillow>=10.2.0
python-dotenv>=1.0.0
google-api-python-client>=2.116.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
requests>=2.31.0
numpy>=1.26.0
pandas>=2.2.0
scipy>=1.12.0
django-storages>=1.14.2
